+++ src/movement_tracking.js
@@ -4,19 +4,19 @@ import {getTokenShape, zip} from "./util.js";
 
 function initTrackingFlag(combatant) {
 	const initialFlag = {passedWaypoints: [], trackedRound: 0};
-	let dragRulerFlag = combatant.flags?.dragRuler;
+	let dragRulerFlag = combatant.data.flags.dragRuler;
 	if (dragRulerFlag) {
 		if (isNaN(dragRulerFlag.trackedRound)) {
 			mergeObject(dragRulerFlag, initialFlag);
 		}
 	}
 	else {
-		combatant.flags.dragRuler = initialFlag;
+		combatant.data.flags.dragRuler = initialFlag;
 	}
 }
 
 function getInitializedCombatant(token, combat) {
-	const combatant = combat.getCombatantByToken(token.data._id);
+	const combatant = combat.getCombatantByToken(token.id);
 	if (!combatant)
 		return undefined;
 	initTrackingFlag(combatant);
@@ -43,14 +43,14 @@ function calculateUpdate(combat, token, rays) {
 		return;
 
 	// Check if we have entered a new round. If so, remove the currently stored path
-	if (combat.data.round > combatant.flags.dragRuler.trackedRound) {
-		combatant.flags.dragRuler.passedWaypoints = [];
-		combatant.flags.dragRuler.trackedRound = combat.data.round;
+	if (combat.data.round > combatant.data.flags.dragRuler.trackedRound) {
+		combatant.data.flags.dragRuler.passedWaypoints = [];
+		combatant.data.flags.dragRuler.trackedRound = combat.data.round;
 	}
 
 	// Add the passed waypoints to the combatant
 	const terrainRulerAvailable = game.modules.get("terrain-ruler")?.active && (!game.modules.get("TerrainLayer")?.active || canvas.grid.type !== CONST.GRID_TYPES.GRIDLESS);
-	const dragRulerFlags = combatant.flags.dragRuler;
+	const dragRulerFlags = combatant.data.flags.dragRuler;
 	const waypoints = dragRulerFlags.passedWaypoints;
 	for (const ray of rays) {
 		// Ignore rays that have the same start and end coordinates
@@ -63,17 +63,17 @@ function calculateUpdate(combat, token, rays) {
 			waypoints.push(ray.A);
 		}
 	}
-	return {_id: combatant._id, dragRulerFlags};
+	return {_id: combatant.id, dragRulerFlags};
 }
 
 export function getMovementHistory(token) {
 	const combat = game.combat;
 	if (!combat)
 		return [];
-	const combatant = combat.getCombatantByToken(token.data._id);
+	const combatant = combat.getCombatantByToken(token.id);
 	if (!combatant)
 		return [];
-	const dragRulerFlags = combatant.flags.dragRuler;
+	const dragRulerFlags = combatant.data.flags.dragRuler;
 	if (!dragRulerFlags)
 		return [];
 	if (combat.data.round > dragRulerFlags.trackedRound)
@@ -82,8 +82,8 @@ export function getMovementHistory(token) {
 }
 
 export async function resetMovementHistory(combat, combatantId) {
-	const combatant = combat.getCombatant(combatantId);
-	const dragRulerFlags = combatant.flags.dragRuler;
+	const combatant = combat.combatants.get(combatantId);
+	const dragRulerFlags = combatant.data.flags.dragRuler;
 	if (!dragRulerFlags)
 		return;
 	dragRulerFlags.passedWaypoints = null;